Refactor spring-ai-model-1.0.0 tests
