+++ instrumentation/spring-ai-model-1.0.0/src/test/java/org/springframework/ai/embedding/AIM_Disabled_ServerSide_EmbeddingModel_InstrumentationTest.java
@@ -33,7 +33,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.springframework.ai.embedding.EmbeddingUtil.buildEmbeddingRequest;
+import static util.EmbeddingUtil.buildEmbeddingRequest;
 
 @Category({ Java8IncompatibleTest.class, Java11IncompatibleTest.class })
 @RunWith(InstrumentationTestRunner.class)
