+++ instrumentation/spring-ai-model-1.0.0/src/test/java/org/springframework/ai/embedding/MockEmbeddingModel.java
@@ -11,7 +11,7 @@ import org.springframework.ai.document.Document;
 
 import java.util.List;
 
-import static org.springframework.ai.embedding.EmbeddingUtil.buildEmbeddingResponse;
+import static util.EmbeddingUtil.buildEmbeddingResponse;
 
 public class MockEmbeddingModel extends AbstractEmbeddingModel {
     @Override
