+++ instrumentation/spring-ai-model-1.0.0/src/test/java/util/EmbeddingUtil.java
@@ -5,12 +5,26 @@
  *
  */
 
-package org.springframework.ai.embedding;
+package util;
 
+import com.newrelic.agent.introspec.Event;
+import llm.embeddings.vendor.Vendor;
 import org.springframework.ai.chat.metadata.DefaultUsage;
+import org.springframework.ai.embedding.Embedding;
+import org.springframework.ai.embedding.EmbeddingOptions;
+import org.springframework.ai.embedding.EmbeddingOptionsBuilder;
+import org.springframework.ai.embedding.EmbeddingRequest;
+import org.springframework.ai.embedding.EmbeddingResponse;
+import org.springframework.ai.embedding.EmbeddingResponseMetadata;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
+
+import static llm.embeddings.events.LlmEvent.LLM_EMBEDDING;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class EmbeddingUtil {
     // request
@@ -45,5 +59,22 @@ public class EmbeddingUtil {
         embeddings.add(embedding);
 
         return new EmbeddingResponse(embeddings, embeddingResponseMetadata);
+    }
+
+    public static void assertLlmEmbeddingAttributes(Event event, String modelId, String requestInput) {
+        assertEquals(LLM_EMBEDDING, event.getType());
+
+        Map<String, Object> attributes = event.getAttributes();
+        assertFalse(((String) attributes.get("id")).isEmpty());
+        assertFalse(((String) attributes.get("request_id")).isEmpty());
+        assertEquals(requestInput, attributes.get("input"));
+        assertEquals(modelId, attributes.get("request.model"));
+        assertEquals(modelId, attributes.get("response.model"));
+        assertEquals(totalTokens, attributes.get("response.usage.total_tokens"));
+        assertEquals(Vendor.VENDOR, attributes.get("vendor"));
+        assertEquals(Vendor.INGEST_SOURCE, attributes.get("ingest_source"));
+        assertTrue(((Float) attributes.get("duration")) >= 0);
+        assertEquals("testPrefix", attributes.get("llm.testPrefix"));
+        assertEquals("conversation-id-value", attributes.get("llm.conversation_id"));
     }
 }
